+++ resources/views/components/jsonld.blade.php
@@ -1,1 +1,1 @@
-<script type="application/ld+json">{!! json_encode($schema ?? $data ?? [], JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES | JSON_PRETTY_PRINT) !!}</script>
+<script type="application/ld+json">{!! app(\Step2dev\LazySeoStructuredData\Services\SchemaService::class)->toJson($schema ?? $data ?? []) !!}</script>